move test settings
